[Performance] Add --optimistic flag that skips DB checks
By default, when running SequenceServer.init i.e. when starting
the web server or runninc CLI commands, it will scan the entire
database_dir entries to find databases that use older (v4) version
or have been created without parse_seqids. If this is the case
it will log a warning.

If all databases are formatted correctly and the user knows it,
running SequenceServer with --optimistic can have a significant
impact on reducing startup times. With vast databse directories
this can be in tens of seconds.

This is because parse_seqids in particular does a lot of IO calls
to scan database directories to determine if seqids metadata files
are missing for each database.