Handle JSON errors as 422
When an issue with a user query or configuration occurs, we were
just raising an exception and the sinatra's error block was
returning 500 error page. When we know the shape of the error
and build a JSON response for it, it's more likely to be a 4**
type of error, so default to 422 for a status in such cases.

This is not ideal, but an improvement when it comes to HTTP status
semantics and also helps manage errors when Sequenceserver is deployed
to a server i.e. passenger can be configured to intercept 500 pages
yet 4** ones would still be handled by the app and display useful
error messages to the user.