Revert "Fix #695 issue"
This reverts commit ecfaa3cde12e970b01ca5cdca8d831b9e42b6f92.

The change had an unintended consequence of disabling the help
modal for advanced features.

I'd argue that https://github.com/wurmlab/sequenceserver/issues/695
is a bit misleading - it is expected for HTML
labels to be clickable. Clicking on an input field label should
focus the input field, which it was doing. I don't think this should
change.

It just happens that the help modal trigger is crammed into the label
too, so it can perhaps get confusing when the two elements next to each other are clickable and react to the same hover - perhaps the latter needs
to change instead?