Add a client side version checker
Pings a backend that notifies if a new version of the gem is
available. It 'caches' the call in a cookie, so it won't be calling
this endpoing more than once a day.

When an update is available, it displays a link to the changelog
reading "New version v${version} is available".