Mock Circos component in tests
A recent change has made it so that Circos graphs are rendered
by default. This has syrfaced an issue that d3 is not set up in
the test suite with jsdom properly and was erring. The tests do
not require Circos graphs to render, so we can mock the component
fully.
R