Add a regression test for report
We're about to refactor the BLAST::Report class, and it does not
seem to have any unit tests. The implementation is a bit hard to
comprehend with generic abbreviated variable names and data accessed
via indexes. To aid the refactoring, create a regression test that
takes a job generated on current master as gold standard and compares
it with the live version. If refactoring changes the outputs, the test
will fail.