Add Cross-Site-Forgery-Request (CSRF) protection
Cross-Site Request Forgery (CSRF) is a type of malicious exploit of a
website where unauthorized commands are transmitted from a user that
the web application trusts. Without the user's knowledge, the attacker
tricks the victim into performing actions on their behalf.
i.e. copy-pasting JS commands to dev tools console or similar.
This vulnerability exploits the trust that a site has in a user's browser,
allowing attackers to perform actions on behalf of the user without
their consent or awareness.

To protect from this attack we're adding a crsf token, that is generated
randomly, stored in session, and added to each form submission/POST
request. Requests fail if the token value submitted does not match
the value in the session, significantly reducing the risk of malicious
or unintentional requests on behalf of the user.

While this is not really necessary for locally running instances or
public instances, folks hosting their own private instances or exposing
locally hosted instances to the internet will benefit from this.