Split off sourcemaps from the minified JS
Configure webpack to compile a separate sourcemaps file instead of
using the default eval inline source maps.

This removes a combined ~3MB from the two JS bundles (the separate
source map files are not loaded by default onto the user's browsers).

Not using eval mode source maps is a bit slower, but also makes it
easier to configure CSP (for the cloud version) with strong defaults