Fix test - reset preselected Databases
It seems that an older commit was added to retain user inputs
when clicking hte back button in the browsers. This was done
by storing the previous job_id in session and restoring the
state from it. However, it has an unintended side effect that
navigating to the root page manually also pre-populates all fields
from the last job. Another fix earlier in this PR has fixed database
pre-select behaviour, so this side effect is now more prominent
and has triggered a test failure.

Keep this behaviour for now, but consider removing in the future
if it annoys us or other users.