Fix modal edge cases
the HTML dialog element allows closing the modal by hitting ESC
key on the keyboard. Doing that was not removing the 'overflow-hidden'
class from document body, keeping the entire page "stuck" and not
scrolling.

Also fix sequence viewer modal with a lot of conent by adding scroll
and max height